+++ src/main/java/com/dmartinez/customer/CustomerRepository.java
@@ -3,8 +3,8 @@ package com.dmartinez.customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository
-        extends JpaRepository<Customer, Long> {
+        extends JpaRepository<Customer, Integer> {
 
     boolean existsCustomerByEmail(String email);
-    boolean existsCustomerById(Long id);
+    boolean existsCustomerById(Integer id);
 }
